refactor(wizard): split model discovery into one discoverer class per provider
ModelDiscovery was 1467 lines: seven providers' listing, filtering,
enrichment and fallback catalogs interleaved in one class. Each provider now
lives in Discovery/<Name>ModelDiscoverer extending a shared base; the facade
keeps its interface, its constructor signature and the connection test, and
shrinks to 241 lines.

Bodies moved verbatim. The behavioural asymmetries between providers are
deliberate and preserved: OpenAI/Anthropic/Gemini/Mistral fall back to static
catalogs while Ollama/OpenRouter/Groq return empty; three providers decode
through the logged JSON path while two decode bare; only OpenAI and Anthropic
sort recommended-first. Fallback-ness now travels on a DiscoveryResult value
returned by each discoverer instead of a flag mutated deep inside the
provider paths — the facade sets its reported state from the result.

Two mechanics carry the tests unchanged. The facade fans its test-only HTTP
seam out to every discoverer, so the ~150 existing tests keep injecting their
stub through the facade and pass untouched. And the trait's dispatch() is
aliased to protected in the base class — flattened private it would be
invisible to the concrete discoverers, and the resulting Error is a
Throwable, so every discovery would have silently caught it and fallen back.
That exact failure produced 40 green-looking fallbacks during the build and
was also the finding of the adversarial plan review.

The SSRF gate is unchanged: every discoverer dispatches through the same
vault client with the same anonymous host pre-gate, and the two seam-less
SSRF tests still prove it fires. ADR-018's dispatch snippet and file list are
updated in place.
